Rename variations declaration page  to package playback.DeclarationPage
